refactoring trainer

+++ src/train.py
@@ -1,4 +1,5 @@
 import os
+import shutil
 import traceback
 import configparser
 
@@ -28,6 +29,105 @@ class Trainer():
         self.log.info("Trainer is ready")
         pass
 
+    def _remove_stored(self, path) -> bool:
+        """
+        Удаление сохраненной ранее модели
+        """
+        if os.path.exists(path):
+            if os.path.isdir(path):
+                shutil.rmtree(path)
+            else:
+                os.remove(path)
+        if os.path.exists(path):
+            self.log.error(f'Can\'t remove {path}')
+            return False
+        return True
+
+    def _calc_watched_matrix(self, grouped) -> bool:
+        """
+        Подготовка матрицы просмотренных фильмов
+        """
+
+        path = "./models/WATCHED"
+        if not self._remove_stored(path):
+            return False
+
+        # возьмём индексы просмотренных фильмов в нужном нам типе, чтобы можно было
+        # умножить их на похожесть пользователя
+        matrix = CoordinateMatrix(grouped.flatMapValues(lambda x: x).map(lambda x: MatrixEntry(x[0], x[1], 1.0)))
+        
+        try:
+            matrix.entries.toDF().write.parquet(path)
+            self.config["MODEL"]["WATCHED_PATH"] = path
+            self.log.info(f"Matrix of watched movies is stored at {path}")
+        except:
+            self.log.error(traceback.format_exc())
+            return False
+        return os.path.exists(path)
+    
+    def _train_tf(self, grouped):
+
+        path = "./models/TF_MODEL"
+        if not self._remove_stored(path):
+            return None
+
+        # чтобы сделать всё правильно, используем DataFrame из-за его встроенного метода columnSimilarities,
+        # позволяющего считать косинусное сходство между колонкам
+        df = grouped.toDF(schema=["user_id", "movie_ids"])
+
+        FEATURES_COUNT = self.config.getint("MODEL", "FEATURES_COUNT", fallback=10000)
+        self.log.info(f'TF-IDF features count = {FEATURES_COUNT}')
+
+        # считаем TF - частоту токенов (фильмов), должна быть 1, т.к. пользователь либо посмотрел, либо не посмотрел фильм
+        hashingTF = HashingTF(inputCol="movie_ids", outputCol="rawFeatures", numFeatures=FEATURES_COUNT)
+        tf_features = hashingTF.transform(df)
+
+        file_path = "./models/TF_MODEL"
+        try:
+            hashingTF.write().overwrite().save(path)
+            self.config["MODEL"]["TF_PATH"] = path
+            self.log.info(f"TF model stored at {path}")
+        except:
+            self.log.error(traceback.format_exc())
+            return None
+        
+        return tf_features
+    
+    def _train_idf(self, tf_features) -> bool:
+
+        path = "./models/IDF_MODEL"
+        if not self._remove_stored(path):
+            return False
+
+        # считаем IDF - здесь уже будут дробные значения, т.к. учёт по пользователям, это и будут фичи
+        idf = IDF(inputCol="rawFeatures", outputCol="features")
+        idfModel = idf.fit(tf_features)
+
+        try:
+            idfModel.write().overwrite().save(path)
+            self.config["MODEL"]["IDF_PATH"] = path
+            self.log.info(f"IDF model stored at {path}")
+        except:
+            self.log.error(traceback.format_exc())
+            return False
+
+        path = "./models/IDF_FEATURES"
+        if not self._remove_stored(path):
+            return False
+
+        idf_features = idfModel.transform(tf_features)
+
+        try:
+            idf_features.write.format("parquet").save(path, mode='overwrite')
+            self.config["MODEL"]["IDF_FEATURES_PATH"] = path
+            self.log.info(f"IDF features stored at {path}")
+        except:
+            self.log.error(traceback.format_exc())
+            return False
+        
+        return True
+
+
     def train_models(self) -> bool:
         try:
             adapter = SparkAdapter()
@@ -39,59 +139,33 @@ class Trainer():
         INPUT_FILENAME = self.config.get("DATA", "INPUT_FILE", fallback="./data/generated.csv")
         self.log.info(f'train data filename = {INPUT_FILENAME}')
 
-        FEATURES_COUNT = self.config.getint("MODEL", "FEATURES_COUNT", fallback=10000)
-        self.log.info(f'TF-IDF features count = {FEATURES_COUNT}')
-
-        # чтение файла как есть
-        raw = sc.textFile(INPUT_FILENAME, adapter.num_partitions)
-        
-        # записи, сгруппированные по user_id
-        grouped = raw.map(lambda x: map(int, x.split())).groupByKey() \
+        # чтение файла, группировка записей по user_id
+        grouped = sc.textFile(INPUT_FILENAME, adapter.num_partitions) \
+            .map(lambda x: map(int, x.split())).groupByKey() \
             .map(lambda x : (x[0], list(x[1])))
-
-        # возьмём индексы просмотренных фильмов в нужном нам типе, чтобы можно было
-        # умножить их на похожесть пользователя
-        all_watched_matrix = CoordinateMatrix(grouped.flatMapValues(lambda x: x).map(lambda x: MatrixEntry(x[0], x[1], 1.0)))
         
-        # чтобы сделать всё правильно, используем DataFrame из-за 
-        # его встроенного метода columnSimilarities, позволяющего
-        # считать косинусное сходство между колонкам
-        df = grouped.toDF(schema=["user_id", "movie_ids"])
+        # расчёт матрицы просмотренных фильмов
+        self.log.info('Calculating matrix of watched movies')
+        if not self._calc_watched_matrix(grouped):
+            return False
+        
+        # обучение TF модели, получение фичей
+        self.log.info('Train TF model')
+        tf_features = self._train_tf(grouped)
+        if tf_features is None:
+            return False
+        
+        # обучение IDF модели
+        self.log.info('Train IDF model')
+        if not self._train_idf(tf_features):
+            return False
 
-        # считаем TF - частоту токенов (фильмов), должна быть 1,
-        #  т.к. пользователь либо посмотрел, либо не посмотрел фильм
-        hashingTF = HashingTF(inputCol="movie_ids", outputCol="rawFeatures", numFeatures=FEATURES_COUNT)
-        tf_features = hashingTF.transform(df)
-
-        model_path = "./models/TF_MODEL"
-        try:
-            hashingTF.save(model_path)
-            self.config["MODEL"]["TF_PATH"] = model_path
-            self.log.info(f"TF model stored at {model_path}")
-        except:
-            self.log.error(traceback.format_exc())
-
-        # считаем IDF - здесь уже будут дробные значения, т.к. учёт по пользователям, это и будут фичи
-        idf = IDF(inputCol="rawFeatures", outputCol="features")
-        idfModel = idf.fit(tf_features)
-
-        model_path = "./models/IDF_MODEL"
-        try:
-            idfModel.save(model_path)
-            self.config["MODEL"]["IDF_PATH"] = model_path
-            self.log.info(f"IDF model stored at {model_path}")
-        except:
-            self.log.error(traceback.format_exc())
-
-
-        idf_features = idfModel.transform(tf_features)
-
-        # сохраняем изменения
+        # сохраняем пути в конфиге
         os.remove(self.config_path)
         with open(self.config_path, 'w') as configfile:
             self.config.write(configfile)
-        pass
-
+        
+        return True
 
 if __name__ == "__main__":
     trainer = Trainer()
